crmapp/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-filter==21.1
django-request-logging==0.7.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
Markdown==3.3.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==1.7.1
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
